misc tiy up
