+++ packages/app-staking/src/library/GenerateNominations/NominationHealth.tsx
@@ -1,35 +1,17 @@
 // Copyright 2026 @polkadot-cloud/polkadot-cloud-apps authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
+import { RetainmentThresholds } from 'consts/retainment'
 import { useHelp } from 'hooks/useHelp'
 import { useNominationHealth } from 'hooks/useNominationHealth'
 import { useTranslation } from 'react-i18next'
+import type { RetainmentStatus } from 'types'
 import { ButtonHelp } from 'ui-buttons'
-import {
-	Badge,
-	CardHeader,
-	Separator,
-	StatusCard,
-	type StatusCardStatus,
-} from 'ui-core/base'
+import { Badge, CardHeader, Separator, StatusCard } from 'ui-core/base'
+import { getRetainmentStatus } from 'utils'
 import type { NominationHealthProps } from './types'
 import { getValidatorsWithRetainment } from './utils'
 import { NominationHealthWrapper } from './Wrapper'
-
-export const RetainmentThresholds = {
-	high: 75,
-	low: 50,
-} as const
-
-const getRetainmentStatus = (rate: number): StatusCardStatus => {
-	if (rate >= RetainmentThresholds.high) {
-		return 'success'
-	}
-	if (rate >= RetainmentThresholds.low) {
-		return 'warning'
-	}
-	return 'danger'
-}
 
 export const NominationHealth = ({
 	isLoading,
@@ -43,35 +25,46 @@ export const NominationHealth = ({
 		validators,
 		retainmentByAddress,
 	)
+	// Get the validators that are below the high retainment threshold and need attention.
 	const validatorsBelowThreshold = validatorsWithRetainment
 		.filter(({ rate }) => rate < RetainmentThresholds.high)
 		.map(({ validator }) => validator)
+
+	// Get the validators that are below the medium retainment threshold and need attention.
 	const hasDangerWarnings =
 		!isLoading &&
-		validatorsWithRetainment.some(({ rate }) => rate < RetainmentThresholds.low)
+		validatorsWithRetainment.some(
+			({ rate }) => rate < RetainmentThresholds.medium,
+		)
+
+	// Use the useNominationHealth hook to manage the nomination health state and actions.
 	useNominationHealth({
 		hasDangerWarnings,
 		onFix,
 		validatorsToFix: validatorsBelowThreshold,
 	})
 
+	// If the data is still loading or there are no validators with retainment data, return null to
+	// avoid rendering the component.
 	if (isLoading || validatorsWithRetainment.length === 0) {
 		return null
 	}
 
+	// Calculate the average retainment rate across all validators with retainment data.
 	const averageRetainment =
 		validatorsWithRetainment.reduce((total, { rate }) => total + rate, 0) /
 		validatorsWithRetainment.length
-	const thresholdWarningStatus: StatusCardStatus =
-		validatorsWithRetainment.some(({ rate }) => rate < RetainmentThresholds.low)
+
+	// Determine the status of the threshold warning based on the retainment rates of the validators.
+	const thresholdWarningStatus: RetainmentStatus =
+		validatorsWithRetainment.some(
+			({ rate }) => rate < RetainmentThresholds.medium,
+		)
 			? 'danger'
 			: 'warning'
+
+	// Determine the overall retainment status based on the average retainment rate.
 	const status = getRetainmentStatus(averageRetainment)
-	const descriptionKey = `averageRetainmentDescription${status[0].toUpperCase()}${status.slice(1)}`
-	const averageRetainmentLabel = `${averageRetainment.toLocaleString(
-		i18n.resolvedLanguage,
-		{ maximumFractionDigits: 1 },
-	)}%`
 
 	return (
 		<NominationHealthWrapper>
@@ -104,11 +97,16 @@ export const NominationHealth = ({
 				status={status}
 				title={
 					<>
-						{t('averageRetainmentScore')}: {averageRetainmentLabel}
+						{t('averageRetainmentScore')}:{' '}
+						{`${averageRetainment.toLocaleString(i18n.resolvedLanguage, {
+							maximumFractionDigits: 1,
+						})}%`}
 					</>
 				}
 			>
-				{t(descriptionKey)}
+				{t(
+					`averageRetainmentDescription${status[0].toUpperCase()}${status.slice(1)}`,
+				)}
 			</StatusCard>
 			{validatorsBelowThreshold.length > 0 && (
 				<StatusCard status={thresholdWarningStatus} role="status">
