+++ packages/app-staking/src/library/NominationList/DetailedItem.tsx
@@ -11,7 +11,6 @@ import { Metrics } from 'library/ListItem/Buttons/Metrics'
 import { Identity } from 'library/ListItem/Labels/Identity'
 import { useNominationStatusData } from 'library/ListItem/Labels/NominationStatus'
 import { RowActionsMenu } from 'library/ValidatorList/RowActionsMenu'
-import { getRateAfterCommission } from 'library/ValidatorList/retainment'
 import { useRetainmentStatsData } from 'library/ValidatorList/useRetainmentStatsData'
 import { useValidatorSelfStake } from 'library/ValidatorList/useValidatorSelfStake'
 import { ValidatorBar } from 'library/ValidatorList/ValidatorBar'
@@ -19,6 +18,7 @@ import { ValidatorCard } from 'library/ValidatorList/ValidatorCard'
 import { ValidatorSummary } from 'library/ValidatorList/ValidatorSummary'
 import { useTranslation } from 'react-i18next'
 import { ListItem } from 'ui-app/ListItem'
+import { getRateAfterCommission } from 'utils'
 import type { ItemProps } from './types'
 
 export const DetailedItem = ({
