+++ packages/app-staking/src/library/ValidatorList/DetailedItem.tsx
@@ -12,12 +12,12 @@ import { Remove } from 'library/ListItem/Buttons/Remove'
 import { ShareLink } from 'library/ListItem/Buttons/ShareLink'
 import type { Validator } from 'types'
 import { ListItem } from 'ui-app/ListItem'
+import { getRateAfterCommission } from 'utils'
 import { FavoriteValidator } from '../ListItem/Buttons/FavoriteValidator'
 import { Select } from '../ListItem/Buttons/Select'
 import { Identity } from '../ListItem/Labels/Identity'
 import { DetailedItemPreloader } from './DetailedItemPreloader'
 import { RowActionsMenu } from './RowActionsMenu'
-import { getRateAfterCommission } from './retainment'
 import type { ItemProps } from './types'
 import { useRetainmentStatsData } from './useRetainmentStatsData'
 import { useValidatorSelfStake } from './useValidatorSelfStake'
