+++ packages/consts/src/retainment.ts
@@ -1,0 +1,8 @@
+// Copyright 2026 @polkadot-cloud/polkadot-cloud-apps authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export const RetainmentThresholds = {
+	high: 75,
+	medium: 50,
+	low: 25,
+} as const
