+++ packages/types/src/validators.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 export type ValidatorStatus = 'waiting' | 'active'
+export type RetainmentStatus = 'success' | 'warning' | 'danger'
 
 export interface Validator {
 	address: string
