+++ packages/utils/package.json
@@ -4,6 +4,9 @@
 	"license": "GPL-3.0-only",
 	"type": "module",
 	"main": "src/index.ts",
+	"exports": {
+		".": "./src/index.ts"
+	},
 	"scripts": {
 		"clear": "rm -rf build tsconfig.tsbuildinfo dist node_modules",
 		"check": "biome check .",
@@ -14,6 +17,7 @@
 		"@w3ux/utils": "3.1.1",
 		"@w3ux/validator-assets": "^1.3.3",
 		"bignumber.js": "^11.1.5",
+		"consts": "workspace:*",
 		"date-fns": "^4.4.0",
 		"dedot": "^1.3.0",
 		"i18next": "^26.3.6",
