+++ packages/utils/src/index.ts
@@ -11,6 +11,7 @@ export * from './observableStore'
 export * from './pages'
 export * from './pools'
 export * from './rate'
+export * from './retainment'
 export * from './size'
 export * from './staking'
 export * from './strings'
