+++ packages/utils/src/retainment.ts
@@ -2,7 +2,11 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import type BigNumber from 'bignumber.js'
+import { RetainmentThresholds } from 'consts/retainment'
+import type { RetainmentStatus } from 'types'
 
+// Check if the self-stake is at or above the hard cap. If the hard cap is not defined or is zero,
+// return false.
 export const isMaxSelfStake = (
 	selfStakePlanck: BigNumber | undefined,
 	hardCapSelfStake: bigint | undefined,
@@ -11,9 +15,13 @@ export const isMaxSelfStake = (
 	hardCapSelfStake > 0n &&
 	selfStakePlanck?.gte(hardCapSelfStake.toString()) === true
 
+// Clamp the retainment rate to be between 0 and 100. If the rate is not a finite number, return
+// 0.
 export const clampRate = (rate: number) =>
 	Number.isFinite(rate) ? Math.min(Math.max(rate, 0), 100) : 0
 
+// Get the retainment rate after commission is applied. If either the rate or commission is not a
+// finite number, return undefined.
 export const getRateAfterCommission = (
 	rate?: number,
 	commission?: number | null,
@@ -25,14 +33,26 @@ export const getRateAfterCommission = (
 		? rate * (1 - commission / 100)
 		: undefined
 
+// Get the retainment status based on the retainment rate.
+export const getRetainmentStatus = (rate: number): RetainmentStatus => {
+	if (rate >= RetainmentThresholds.high) {
+		return 'success'
+	}
+	if (rate >= RetainmentThresholds.medium) {
+		return 'warning'
+	}
+	return 'danger'
+}
+
+// Get the color associated with the retainment rate.
 export const getRateColor = (rate: number): string => {
-	if (rate >= 75) {
+	if (rate >= RetainmentThresholds.high) {
 		return 'var(--status-success)'
 	}
-	if (rate >= 50) {
+	if (rate >= RetainmentThresholds.medium) {
 		return 'var(--status-warning)'
 	}
-	if (rate >= 25) {
+	if (rate >= RetainmentThresholds.low) {
 		return 'color-mix(in srgb, var(--status-warning) 55%, var(--status-danger))'
 	}
 	return 'var(--status-danger)'
